hola

+++ HolaSDL/Vector2D.cpp
@@ -4,7 +4,7 @@
 typedef unsigned int uint;
 
 Vector2D::Vector2D() : x(), y() {}
-Vector2D::Vector2D(uint x, uint y) : x(x), y(y) {}
+Vector2D::Vector2D(double x, double y) : x(x), y(y) {}
 
 double Vector2D::getX() const {
 	return x;
